Create README.md
Abstract: 
This work presents a smart compactible integration of EV with the distribution grid assuring a quality power. The proposed model of an electric vehicle recharging station is suitable for the fast DC charging of multiple electric vehicles.
In this study, work is done for a level 3 off-board EV charging infrastructure that can employ multiple EVs which will alleviate the down-time required for vehicle charging. There is a master control to deal with the power exchange between the AC and DC bus (AC-DC Converter Control). The control of individual EV is de-centralized (DC-DC Converter Control). The work looks into two different aspects of EV integration to grid a) Grid to Vehicle mode-G2V (EV charging), b) Vehicle to Grid Mode-V2G (EV discharging Mode). 
Design guidelines and modeling are explained in an educational way to support implementation in Matlab/Simulink. Simulations are performed to illustrate the behavior of the station. The results show the feasibility of the model proposed and the capability of the control system for fast DC charging and also vehicle-to-grid.

Index Terms: Electric Vehicle (EV), Vehicle-to-Grid (V2G), MATLAB/Simulink, Constant Current (CC) and Constant Voltage (CV).

References:
1. Arnaldo Arancibia and Kai Strunz, “Modeling of an electric vehicle charging station for fast DC charging,” in IEEE International Electric Vehicle Conference, 2012, pp. 1-6. 

2. Polly Thomas, Fossy Mary Chacko, “Electric Vehicle Integration to Distribution Grid Ensuring Quality Power Exchange”, International Journal of Advanced Research in Electrical, Electronics and Instrumentation Engineering (IJAREEIE), Vol. 2, Special Issue, Dec 2013, pp. 403 – 412.